changed ganache container

+++ utils/migrate.sh
@@ -1,6 +1,9 @@
 cd apps/tokens/solidity/
-rm -R build/contracts
-truffle compile
-truffle migrate --reset
-rm -R ../../../static/json/
+if [ -d "build/contracts" ] ; then
+    rm -R build/contracts
+fi
+
+if [ -d "../../../static/json/" ] ; then
+    rm -R ../../../static/json/
+fi
 cp -R build/contracts ../../../static/json/
